#186 ユーザ退会

+++ app/Http/Controllers/UsersController.php
@@ -15,10 +15,11 @@ class UsersController extends Controller
     public function destroy($id)
     {
         $user = User::findOrFail($id);
-        // if (\Auth::id() === $user->id) {
+        if (\Auth::id() === $user->id) {
             $user->delete();
             
             return redirect('/');
+        }                
     }
 
     public function show($id)
